Ulternate movement sceme for spacecoil
I was experimenting with an alternative control scheme in which the player controlled their g.u.n. with wasd and the grappling hook was mapped to mouse.  The player could switch to laser mode by holding space. However, I came to the conclusion that the previous control scheme was better.